+++ CHANGES.md
@@ -1,6 +1,9 @@
 
 # Unreleased
 
+* add: support for `bands` options in `STACReader._get_asset_info` methods to select band indexes based on band metadata
+* add: optional `indexes`, `expression` and `bands` keys to `AssetWithOptions` typedDict definition
+
 # 9.0.0b1 (2026-02-22)
 
 * add `AssetType = str | AssetWithOptions`, with `AssetWithOptions` beeing a typed dict with required `name` key